@@ -8,8 +8,11 @@
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
+    "axios": "^0.24.0",
+    "bootstrap": "5.1.3",
     "core-js": "^3.6.5",
-    "vue": "^2.6.11"
+    "vee-validate": "^4.5.7",
+    "vue": "^3.2.26"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "~4.5.0",
