@@ -1,8 +1,9 @@
-import Vue from 'vue'
+import { createApp } from 'vue'
 import App from './App.vue'
+import "bootstrap/dist/css/bootstrap.min.css";
+import './assets/css/main.css'
+import './assets/css/media.css'
 
-Vue.config.productionTip = false
 
-new Vue({
-  render: h => h(App),
-}).$mount('#app')
+
+createApp(App).mount('#app')
